Fix warning.
git-svn-id: svn://localhost/ardour2/branches/3.0@10825 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/audio_clock.cc
@@ -1808,7 +1808,7 @@ AudioClock::timecode_validate_edit (const string& str)
 		return false;
 	}
 
-	if (TC.frames > (long)rint(_session->timecode_frames_per_second()) - 1) {
+	if (TC.frames > (uint32_t) rint (_session->timecode_frames_per_second()) - 1) {
 		return false;
 	}
